requirements.txt
flask>=3.0
playwright>=1.40
requests>=2.31
google-generativeai>=0.4
python-dotenv>=1.0
